1.suppor包升级到androidx 2.依赖包版本升级 3.添加MMKV的支持 4.MVP支持ViewModel

+++ arad/src/main/java/com/beanu/arad/utils/fastblur/BlurBuilder.java
@@ -48,8 +48,8 @@ public class BlurBuilder {
             int width = Math.round(image.getWidth() * BITMAP_SCALE);
             int height = Math.round(image.getHeight() * BITMAP_SCALE);
 
-            overlay = Bitmap.createScaledBitmap(image, (int) (width),
-                    (int) (height), false);
+            overlay = Bitmap.createScaledBitmap(image, width,
+                    height, false);
 
             overlay = FastBlurUtil.doBlur(overlay, (int) BLUR_RADIUS, true);
         } catch (Exception e) {
